[US403584] UCP service accounts & endpoints to global (optimize site design)
This PS begins the efforts of optimizing the site design to account for
a less than desired current state. Here, we'll move UCP endpoints and
service accounts up to global, as they should exist for each site
(UCP is the foundation we're building on top of) and should almost always
be the same for each site.

In addition, this PS also clears up a few linting issues that I
encountered when running Pegleg.

Change-Id: I48095a4b2e4c836f02fe05d3cf111c06f920a576