[US403584] Coredns cleanup (optimize site design)
*This PS removes the coredns chart from the atl-lab1 site, as
it was identical to global. In addition, the global version was
already set to non-abstract, so there were previously 2 concrete
coredns documents in the manifests. I adjusted the global metadata
name, so that pre-existing manifests will continue to pick the
chart up.
*I adjusted the style of the substitution patterns by removing
single quotes and circle brackets from them. Capture groups were
the original intent here, but since that's not implemented, this
should make things cleaner.

Change-Id: I4d54a3d0378195938e995a1caeed8d722e30e464