Add Schemas for Shipyard Deployment files
Adds the DeploymentStrategy Schema. Modifies the
DeploymentConfiguration schema to no longer require the deployment
strategy be a value in the enum list (instead it's a string now that
specifies the DeploymentStrategy document name). This change also
reomves the deployment strategy field from the deployment configuration
document, which will allow the Shipyard validator to not require a
deployment Strategy document names `all-at-once`.

Subsequent changes to Shipyard will allow for this same arrangement, and
will assume an `all-at-once` strategy if no value is specified for
deployment strategy, but will also use a deployment strategy if it is
specified.

Without this change, the current Shipyard code will fail validation
during commit of configuration documents. This chaange is prompted by
the move to newer images that are in-between the introduction of the
validation of deployment strategy (currently implemented) and the (coming
soon) implementation of using the deployment strategy in the workflow
during deploy-site and update-site.

Change-Id: I8fcb73e5f0d234ea7aded1fc4ac3f61c3e27f06c