Update patch set 5
Patch Set 5: Code-Review+2

Talked to Ruslan offline about this. The shipyard reference in the deployment configuration yaml itself references an old version of shipyard for which src/bin/shipyard_airflow/schemas doesn't exist -- as the last time TreatureMap was updated was April 2, which predates the Shipyard repository directory layout change. But since the 'id' field in the schema is metadata this is a safe change so I'm OK with approving it.

Patch-set: 5
Reviewer: Felipe Monteiro <1011942@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2